feat: implement temperature, top-k, and top-p filtering functions in utilities
